set type uint32 uid,gid item in windows

+++ pkg/backupapi/item_windows.go
@@ -13,8 +13,8 @@ func ItemLocal(fi fs.FileInfo) (time.Time, time.Time, time.Time, uint32, uint32)
 		ctimeLocal = time.Unix(0, stat.LastAccessTime.Nanoseconds()).UTC()
 		mtimeLocal = time.Unix(0, stat.LastAccessTime.Nanoseconds()).UTC()
 	}
-	uid := 0
-	gid := 0
+	uid := uint32(0)
+	gid := uint32(0)
 
 	return atimeLocal, ctimeLocal, mtimeLocal, uid, gid
 }
